backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dj-database-url==3.1.2
dj-rest-auth==7.1.1
Django==5.1.15
django-allauth==65.14.3
django-cors-headers==4.9.0
django-jazzmin==3.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
oauthlib==3.3.1
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
python-decouple==3.8
reportlab==4.4.10
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
